Refactor files in middleware/transport
