+++ src/middleware/application/multi_thread.h
@@ -27,7 +27,7 @@
 
 #include "goby/exception.h"
 #include "goby/middleware/application/interface.h"
-#include "goby/middleware/thread.h"
+#include "goby/middleware/application/thread.h"
 #include "goby/middleware/transport/interprocess.h"
 #include "goby/middleware/transport/interthread.h"
 #include "goby/middleware/transport/intervehicle.h"
