+++ src/middleware/application/single_thread.h
@@ -26,7 +26,7 @@
 #include <boost/units/systems/si.hpp>
 
 #include "goby/middleware/application/interface.h"
-#include "goby/middleware/thread.h"
+#include "goby/middleware/application/thread.h"
 #include "goby/middleware/transport/interprocess.h"
 #include "goby/middleware/transport/intervehicle.h"
