+++ src/middleware/transport/interface.h
@@ -34,8 +34,8 @@
 #include "goby/exception.h"
 #include "goby/middleware/protobuf/intervehicle.pb.h"
 #include "goby/middleware/protobuf/transporter_config.pb.h"
-#include "goby/middleware/publisher.h"
-#include "goby/middleware/subscriber.h"
+#include "goby/middleware/transport/publisher.h"
+#include "goby/middleware/transport/subscriber.h"
 #include "goby/util/debug_logger.h"
 
 namespace goby
